Extract era module from era supervisor.
